разбивает утилитарный модуль на несколько модулей
