+++ src/utils/common.js
@@ -1,0 +1,32 @@
+/**
+ * Функция из интернета по генерации случайного числа из диапазона
+ * Источник - https://github.com/you-dont-need/You-Dont-Need-Lodash-Underscore#_random
+ * @param a - минимальное значение
+ * @param b - максимальное значение
+ * @returns {number}
+ */
+export const getRandomInteger = (a = 0, b = 1) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+
+  return Math.floor(lower + Math.random() * (upper - lower + 1));
+};
+
+/**
+ * Функция, возвращающая случайное целое число из переданного диапазона включительно.
+ * Результат: целое число из диапазона "от...до".
+ * Источник: https://developer.mozilla.org/ru/docs/Web/JavaScript/Reference/Global_Objects/Math/random
+ *
+ * @param min - минимальное значекние диапозона
+ * @param max - максимальное значение диапозона
+ */
+export const getRandomIntInclusive = (min, max) => {
+  const minNumber = Math.ceil(min);
+  const maxNumber = Math.floor(max);
+  if (minNumber >= maxNumber) {
+    return 'Введены неверные значения';
+  }
+
+  // Максимум и минимум включаются
+  return Math.floor(Math.random() * (maxNumber - minNumber + 1)) + minNumber;
+};
